// aah_apr_front_end-master/types/Types.ts
export type Evaluation = {

  participant: string;
  selected: string;
};